requirements.txt
line-bot-sdk==2.4.2
flask==2.2.3
requests==2.28.2

pip~=20.2.4
wheel~=0.35.1
setuptools~=50.3.2
future~=0.18.3
pygsheets~=2.0.6
numpy~=1.21.6
popdf~=0.0.3